75 Associations - Interim Commit 8 - Move Descriptor code into a Mixin
